+++ src/params.rs
@@ -1,22 +1,14 @@
 use crate::detect_anoms::detect_anoms;
+use crate::Error;
 
+#[derive(Debug)]
 pub enum Direction {
     Positive,
     Negative,
     Both,
 }
 
-// TODO remove in 0.2.0
-impl Into<String> for Direction {
-    fn into(self) -> String {
-        match self {
-            Direction::Positive => "pos".to_string(),
-            Direction::Negative => "neg".to_string(),
-            _ => "both".to_string(),
-        }
-    }
-}
-
+#[derive(Debug)]
 pub struct AnomalyDetectionParams {
     alpha: f32,
     max_anoms: f32,
@@ -33,6 +25,7 @@ pub fn params() -> AnomalyDetectionParams {
     }
 }
 
+#[derive(Debug)]
 pub struct AnomalyDetectionResult {
     anomalies: Vec<usize>,
 }
@@ -54,18 +47,8 @@ impl AnomalyDetectionParams {
         self
     }
 
-    // TODO only support enum in 0.2.0
-    pub fn direction<S: Into<String>>(&mut self, value: S) -> &mut Self {
-        let direction = value.into();
-        if direction == "pos".to_string() {
-            self.direction = Direction::Positive;
-        } else if direction == "neg".to_string() {
-            self.direction = Direction::Negative;
-        } else if direction == "both".to_string() {
-            self.direction = Direction::Both;
-        } else {
-            panic!("direction must be pos, neg, or both")
-        }
+    pub fn direction(&mut self, value: Direction) -> &mut Self {
+        self.direction = value;
         self
     }
 
@@ -74,16 +57,15 @@ impl AnomalyDetectionParams {
         self
     }
 
-    // TODO return Result in 0.2.0
-    pub fn fit(&self, series: &[f32], period: usize) -> AnomalyDetectionResult {
+    pub fn fit(&self, series: &[f32], period: usize) -> Result<AnomalyDetectionResult, Error> {
         let (one_tail, upper_tail) = match self.direction {
             Direction::Positive => (true, true),
             Direction::Negative => (true, false),
             Direction::Both => (false, true),
         };
 
-        AnomalyDetectionResult {
-            anomalies: detect_anoms(series, period, self.max_anoms, self.alpha, one_tail, upper_tail, self.verbose).unwrap_or_else(|e| panic!("{}", e)),
-        }
+        Ok(AnomalyDetectionResult {
+            anomalies: detect_anoms(series, period, self.max_anoms, self.alpha, one_tail, upper_tail, self.verbose)?,
+        })
     }
 }