+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0 (unreleased)
+
+- Changed `fit` to return `Result`
+- Dropped support for `String` direction
+
 ## 0.1.2 (2021-10-18)
 
 - Added `verbose` option